# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Participium API - Citizen Registration & Municipality Administration
  description: |
    API for citizen registration, authentication (PT01), municipality user management (PT02) and municipality users role assignment (PT03).
    Uses session-based authentication with HTTP-only cookies.

    ## Municipality User Management
    System administrators can create, update and delete internal municipality user accounts.
    The system must have at least one administrator account pre-configured in the database.

    **Role Categories:**

    **MUNICIPALITY_ROLES** (all internal municipality staff):
    - ADMINISTRATOR: System administrator with full privileges
    - PUBLIC_RELATIONS: Public relations officer (only role that can approve/reject reports)
    - CULTURE_EVENTS_TOURISM_SPORTS: Culture, events, tourism and sports
    - LOCAL_PUBLIC_SERVICES: Local public services
    - EDUCATION_SERVICES: Education services
    - PUBLIC_RESIDENTIAL_HOUSING: Public/residential housing services
    - INFORMATION_SYSTEMS: Information systems/IT
    - MUNICIPAL_BUILDING_MAINTENANCE: Municipal building maintenance
    - PRIVATE_BUILDINGS: Private buildings oversight
    - INFRASTRUCTURES: Road, bridges and general infrastructures
    - GREENSPACES_AND_ANIMAL_PROTECTION: Parks, green areas and animal protection
    - WASTE_MANAGEMENT: Waste collection and management
    - ROAD_MAINTENANCE: Road maintenance
...

...

...
    
    **External roles:**
...
  version: 1.1.0
  contact:
...
servers:
...
paths:
...
  /reports:
    post:
      tags:
        - Reports
      summary: Create a new report with geolocation
      description: |
        Create a new citizen report tied to a geographic location within the municipality of Turin.
        Use an interactive map to select the location. Coordinates must be in decimal degrees (WGS84).
        All coordinates must be within Turin municipality boundaries.
      operationId: createReport
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ReportRequest"
            encoding:
              photos:
                contentType: image/jpeg, image/png
      responses:
        "201":
          description: Report created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  report:
                    $ref: "#/components/schemas/ReportResponse"
              example:
                message: "Report created successfully"
                report:
                  id: 1024
                  title: "Pothole on Via Roma"
                  description: "Large pothole near the pedestrian crossing, causing risk for cyclists."
                  category: ROADS_URBAN_FURNISHINGS
                  latitude: 45.070339
                  longitude: 7.686864
                  status: PENDING_APPROVAL
                  user:
                    id: 201
                    firstName: Mario
                    lastName: Rossi
                    email: mario.rossi@example.com
                    role: CITIZEN
                    telegramUsername: null
                    emailNotificationsEnabled: true
                  assignedOfficer: null
                  messages:
                    - id: 1
                      content: "Initial report created by user"
                      createdAt: "2025-11-11T10:12:34Z"
                      senderId: 201
                  rejectedReason: null
                  photos:
                    - id: 301
                      url: "https://cdn.example.com/reports/1024/photo1.jpg"
                      filename: "photo1.jpg"
                  createdAt: "2025-11-11T10:12:34Z"
                  updatedAt: "2025-11-11T10:12:34Z"
                  
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 400
                error: "BadRequest"
                message: "Missing required fields"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "User not logged in"
        "403":
          description: Insufficient permissions (only citizens can create reports)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 403
                error: "Forbidden"
                message: "Only citizens can create reports"
        "413":
          description: Photo file(s) too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 413
                error: "PayloadTooLarge"
                message: "One or more photos exceed the maximum file size"
        "422":
          description: Invalid coordinates (outside Turin municipality boundaries)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 422
                error: "UnprocessableEntity"
                message: "Coordinates are outside Turin municipality boundaries"
        "500":
          description: Unable to create report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                error: "InternalServerError"
                message: "Unable to create report"
    get:
      tags:
        - Reports
      summary: Get all approved reports for map visualization
      description: |
        Retrieve all approved reports for display on an interactive map. This endpoint returns only reports 
        that have been approved by public relations officers and are in one of the following statuses:
        - ASSIGNED (approved and assigned to technical staff)
        - EXTERNAL_ASSIGNED (assigned to external company/maintainer)
        - IN_PROGRESS (being worked on)
        - RESOLVED (completed)

        Reports with status PENDING_APPROVAL or REJECTED are NOT included in the response.

        This endpoint is designed for citizen-facing map views where users can see all active and resolved 
        issues in their area. The response includes reporter information (name or "anonymous" flag), location 
        coordinates, and other details needed for map markers and popups.

        Optionally filter by a bounding box using the `bbox` query parameter in the format 
        `minLon,minLat,maxLon,maxLat` (decimal degrees, WGS84). This is useful for map views where the 
        frontend requests only reports within the currently visible map area.

        The `category` query parameter is also optional and can be combined with `bbox` to request
        reports of a given category inside the visible map area (e.g. `?bbox=7.66,45.06,7.70,45.08&category=PUBLIC_LIGHTING`).
      operationId: listReports
      parameters:
        - name: category
          in: query
          required: false
          description: Filter reports by category
          schema:
            $ref: "#/components/schemas/ReportCategory"
          example: ROADS_URBAN_FURNISHINGS
      responses:
        "200":
          description: Array of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportResponse"
              example:
                - id: 1024
                  title: "Pothole on Via Roma"
                  description: "Large pothole near the pedestrian crossing, causing risk for cyclists."
                  category: ROADS_URBAN_FURNISHINGS
                  latitude: 45.070339
                  longitude: 7.686864
                  status: ASSIGNED
                  user:
                    id: 201
                    firstName: Mario
                    lastName: Rossi
                    email: mario.rossi@example.com
                    role: CITIZEN
                    telegramUsername: null
                    emailNotificationsEnabled: true
                  assignedOfficer:
                    id: 305
                    firstName: Luca
                    lastName: Verdi
                    email: luca.verdi@municipality.com
                    role: TECHNICAL_OFFICE
                  messages:
                    - id: 1
                      content: "Initial report created by user"
                      createdAt: "2025-11-11T10:12:34Z"
                      senderId: 201
                      senderRole: CITIZEN
                    - id: 2
                      content: "Report approved by public relations officer"
                      createdAt: "2025-11-11T11:15:00Z"
                      senderId: 123
                      senderRole: PUBLIC_RELATIONS
                  rejectedReason: null
                  photos:
                    - id: 301
                      url: "https://cdn.example.com/reports/1024/photo1.jpg"
                      filename: "photo1.jpg"
                  createdAt: "2025-11-11T10:12:34Z"
                  updatedAt: "2025-11-11T10:12:34Z"
                  
                - id: 1025
                  title: "Broken street light"
                  description: "Street light not working"
                  category: PUBLIC_LIGHTING
                  latitude: 45.071000
                  longitude: 7.687500
                  status: RESOLVED
                  user:
                    id: 202
                    firstName: Maria
                    lastName: Bianchi
                    email: maria.bianchi@example.com
                    role: CITIZEN
                    telegramUsername: null
                    emailNotificationsEnabled: true
                  assignedOfficer:
                    id: 306
                    firstName: Paolo
                    lastName: Neri
                    email: paolo.neri@municipality.com
                    role: TECHNICAL_OFFICE
                  messages: []
                  rejectedReason: null
                  photos: []
                  createdAt: "2025-11-10T09:00:00Z"
                  updatedAt: "2025-11-12T08:00:00Z"
                  
        "400":
          description: Invalid category parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 400
                error: "BadRequest"
                message: "Invalid category value"
        "500":
          description: Unable to retrieve reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                error: "InternalServerError"
                message: "Error during report retrieval"
  /reports/pending:
    get:
      tags:
        - Reports
      summary: Get reports pending approval
      description: |
        Retrieve all reports with status PENDING for review by PUBLIC_RELATIONS officers.
        This is a specialized endpoint for PUBLIC_RELATIONS staff only to review and process citizen reports.
        Only PUBLIC_RELATIONS can approve, reject, or assign reports to technical staff.
        Technical staff and external maintainers cannot access pending reports.
      operationId: getPendingReports
      security:
        - SessionAuth: []
      responses:
        "200":
          description: List of pending reports retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportResponse"
              example:
                - id: 1024
                  title: "Pothole on Via Roma"
                  description: "Large pothole near the pedestrian crossing, causing risk for cyclists."
                  category: ROADS_URBAN_FURNISHINGS
                  latitude: 45.070339
                  longitude: 7.686864
                  status: PENDING_APPROVAL
                  user:
                    id: 201
                    firstName: Mario
                    lastName: Rossi
                    email: mario.rossi@example.com
                    role: CITIZEN
                    telegramUsername: null
                    emailNotificationsEnabled: true
                  assignedOfficer: null
                  messages:
                    - id: 1
                      content: "Initial report created by user"
                      createdAt: "2025-11-11T10:12:34Z"
                      senderId: 201
                  rejectedReason: null
                  photos:
                    - id: 301
                      url: "https://cdn.example.com/reports/1024/photo1.jpg"
                      filename: "photo1.jpg"
                  createdAt: "2025-11-11T10:12:34Z"
                  updatedAt: "2025-11-11T10:12:34Z"
                  
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "Authentication required"
        "403":
          description: Insufficient permissions (PUBLIC_RELATIONS role required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 403
                error: "Forbidden"
                message: "Public relations officer privileges required"
        "500":
          description: Unable to retrieve pending reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                error: "InternalServerError"
                message: "Failed to retrieve pending reports"
  /reports/{reportId}/assignable-technicals:
    get:
      tags:
        - Reports
      summary: List technicals available for this report's category
      description: |
        Return the list of municipality technical users (TECHNICAL_ROLES) that are valid assignees for the given report.
        The backend filters available technicals according to the report's category. This endpoint is
        intended for use by PUBLIC_RELATIONS officers only when approving a report and choosing the responsible technical.
        External maintainers cannot access this endpoint as they cannot approve reports.
      operationId: getAssignableTechnicals
      security:
        - SessionAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          description: ID of the report
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Array of municipality technical users suitable for assignment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MunicipalityUserResponse"
              example:
                - id: 305
                  firstName: Luca
                  lastName: Verdi
                  email: luca.verdi@municipality.com
                  role: MUNICIPAL_BUILDING_MAINTENANCE
                - id: 306
                  firstName: Paolo
                  lastName: Neri
                  email: paolo.neri@municipality.com
                  role: ROAD_MAINTENANCE
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "Authentication required"
        "403":
          description: Insufficient permissions (PUBLIC_RELATIONS role required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 403
                error: "Forbidden"
                message: "Public relations officer privileges required"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 404
                error: "NotFound"
                message: "Report not found"
        "500":
          description: Unable to retrieve technicals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                error: "InternalServerError"
                message: "Failed to retrieve assignable technicals"
...
  /reports/{reportId}/reject:
    post:
      tags:
        - Reports
      summary: Reject a report
      description: |
        Reject a report with a mandatory explanation. Only public relations officers can perform this operation.
        The report status changes from PENDING_APPROVAL to REJECTED.
        A rejection reason must be provided in the request body.
      operationId: rejectReport
      security:
        - SessionAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          description: ID of the report to reject
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectReportRequest"
            example:
              reason: "The reported issue does not fall under municipal jurisdiction"
      responses:
        "200":
          description: Report rejected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  report:
                    $ref: "#/components/schemas/ReportResponse"
              example:
                message: "Report rejected successfully"
                report:
                  id: 1024
                  title: "Pothole on Via Roma"
                  description: "Large pothole near the pedestrian crossing, causing risk for cyclists."
                  category: ROADS_URBAN_FURNISHINGS
                  latitude: 45.070339
                  longitude: 7.686864
                  status: REJECTED
                  user:
                    id: 201
                    firstName: Mario
                    lastName: Rossi
                    email: mario.rossi@example.com
                    role: CITIZEN
                    telegramUsername: null
                    emailNotificationsEnabled: true
                  assignedOfficer: null
                  messages:
                    - id: 1
                      content: "Initial report created by user"
                      createdAt: "2025-11-11T10:12:34Z"
                      senderId: 201
                    - id: 2
                      content: "Report rejected by public relations officer"
                      createdAt: "2025-11-11T11:15:00Z"
                      senderId: 123
                  rejectedReason: "The reported issue does not fall under municipal jurisdiction"
                  photos:
                    - id: 301
                      url: "https://cdn.example.com/reports/1024/photo1.jpg"
                      filename: "photo1.jpg"
                  createdAt: "2025-11-11T10:12:34Z"
                  updatedAt: "2025-11-11T11:15:00Z"
                  
        "400":
          description: Invalid report ID, missing rejection reason, or report not in PENDING_APPROVAL status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 400
                error: "BadRequest"
                message: "Missing required field: reason"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "Authentication required"
        "403":
          description: Insufficient permissions (PUBLIC_RELATIONS role required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 403
                error: "Forbidden"
                message: "Public relations officer privileges required"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 404
                error: "NotFound"
                message: "Report not found"
        "409":
          description: Report already processed by another officer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 409
                error: "Conflict"
                message: "Report has already been processed"
        "422":
          description: Invalid rejection reason (empty, too long, or inappropriate content)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 422
                error: "UnprocessableEntity"
                message: "Rejection reason must be between 1 and 500 characters"
        "500":
          description: Unable to reject report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                error: "InternalServerError"
                message: "Failed to reject report"
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: |
        Session-based authentication using HTTP-only cookies with Passport.js.
        The session cookie is automatically sent by the browser with each request.
        Frontend must use 'credentials: include' in fetch/axios requests.
  schemas:
...
    Latitude:
      type: string
      description: Latitude in decimal degrees (WGS84)
      example: 45.070339
    Longitude:
      type: string
      description: Longitude in decimal degrees (WGS84)
      example: 7.686864
    ReportRequest:
      type: object
      required:
        - title
        - description
        - category
        - latitude
        - longitude
        - photos
      properties:
        title:
          type: string
          description: Short title for the report
          example: "Pothole on Via Roma"
        description:
          type: string
          description: Detailed description of the issue
          example: "Large pothole near the pedestrian crossing, causing risk for cyclists."
        category:
          $ref: "#/components/schemas/ReportCategory"
        latitude:
          $ref: "#/components/schemas/Latitude"
        longitude:
          $ref: "#/components/schemas/Longitude"
        address:
          type: string
          nullable: true
          description: Optional human-readable address for the location
          example: "Via Roma 1, Torino"
        photos:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: string
            format: binary
          description: Up to 3 photo files (image/jpeg, image/png)
...